// src/enums/venueNames.ts
export enum VenueNames {
  TYNESIDE_CINEMA = "Tyneside Cinema",
  GATESHEAD_VUE = "Gateshead Vue",
  CRAMLINGTON_VUE = "Cramlington Vue",
  ODEON_SILVERLINK = "Odeon Silverlink",
  ODEON_METROCENTRE = "Odeon Metrocentre",
  EVERYMAN_NEWCASTLE = "Everyman Newcastle",
  CINEWORLD_NEWCASTLE = "Cineworld Newcastle",
  CINEWORLD_BOLDON = "Cineworld Boldon"
}